+++ packages/abstract-provider/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "abstract-provider/5.0.0-beta.138";
+export declare const version = "abstract-provider/5.0.0-beta.139";