+++ packages/experimental/lib/nonce-manager.d.ts
@@ -1,10 +1,10 @@
 import { ethers } from "ethers";
 export declare class NonceManager extends ethers.Signer {
     readonly signer: ethers.Signer;
-    readonly provider: ethers.providers.Provider;
     _initialPromise: Promise<number>;
     _deltaCount: number;
     constructor(signer: ethers.Signer);
+    get provider(): ethers.providers.Provider;
     connect(provider: ethers.providers.Provider): NonceManager;
     getAddress(): Promise<string>;
     getTransactionCount(blockTag?: ethers.providers.BlockTag): Promise<number>;